refactor(terminal): 拆出剪贴板选区与 xterm/WebSocket 会话主链路 hook
- 新增 components/terminal/：useTerminalClipboard（复制/粘贴/划选复制与 mouseup 丢失防呆）、useTerminalSession（xterm+WS 建立、智能写入、命令记录、TUI 备用屏、resize/重连/状态提示）、terminalLocalEchoFilter（本地回显预测过滤状态机）、terminalKeyEventHandler（自定义快捷键处理器工厂）
- Terminal.tsx 3391 -> 2365 行
