+++ frontend/src/components/Terminal.tsx
@@ -29,6 +29,8 @@ import { useTerminalLinkUnderlines } from './terminal/useTerminalLinkUnderlines.
 import { useTerminalGutter } from './terminal/useTerminalGutter.ts';
 import { useTerminalTheme } from './terminal/useTerminalTheme.ts';
 import { useTerminalSettingsEvents } from './terminal/useTerminalSettingsEvents.ts';
+import { useTerminalClipboard } from './terminal/useTerminalClipboard.ts';
+import { useTerminalSession } from './terminal/useTerminalSession.ts';
 import type { TerminalProps } from './terminal/terminalTypes.ts';
 import { ToggleSwitch } from './settings/SharedComponents.tsx';
 import type { QuickCommandsHandle } from './QuickCommands.tsx';
@@ -68,9 +70,6 @@ export default function Terminal({
   const searchAddonRef = useRef<SearchAddon | null>(null);
   const termSearchInputRef = useRef<HTMLInputElement | null>(null);
   const wsRef          = useRef<WebSocket | null>(null);
-  const connectWebSocketRef = useRef<(() => void) | null>(null);
-  const statusRef      = useRef(status);
-  useEffect(() => { statusRef.current = status; }, [status]);
   const serverIdRef    = useRef(serverId);
   serverIdRef.current  = serverId;
   const [contextMenu, setContextMenu]         = useState<{ x: number; y: number; source: 'terminal' | 'input' } | null>(null);
@@ -206,22 +205,17 @@ export default function Terminal({
   const terminalRightClickPasteModeRef = useRef(localStorage.getItem('terminalRightClickPasteMode') === 'always' ? 'always' : 'empty');
   const terminalLeftClickCopyOnSelectionRef = useRef(localStorage.getItem('terminalLeftClickCopyOnSelection') === 'true');
   const terminalLeftClickCopyOnSelectionModeRef = useRef(localStorage.getItem('terminalLeftClickCopyOnSelectionMode') === 'mouseup' ? 'mouseup' : 'click');
-  const terminalMouseDownSelectionRef = useRef<{ mode: 'mouseup' | 'click'; startClientX: number; startClientY: number; text?: string } | null>(null);
-  const isTerminalPointerDownRef = useRef(false);
-  // macOS WKWebView / 系统手势可能吞掉 mouseup，导致 xterm 拖选状态机卡死（此后划动指针 = 持续划选）。
-  // 主动向 document 派发合成 mouseup 闭合状态机：xterm 的选区收尾监听挂在 document 上且不校验 isTrusted；
-  // 事件坐标仅 altClickMovesCursor 分支使用（合成事件 altKey=false 不会进入），无指针信息时传 0 即可。
-  const dispatchSyntheticTerminalMouseUp = useCallback((clientX = 0, clientY = 0) => {
-    document.dispatchEvent(new MouseEvent('mouseup', {
-      bubbles: true,
-      cancelable: true,
-      view: window,
-      clientX,
-      clientY,
-      buttons: 0,
-      button: 0,
-    }));
-  }, []);
+  const { isTerminalPointerDownRef, dispatchSyntheticTerminalMouseUp, pasteClipboardToTerminal, pasteTerminalSelectionToTerminal, handleTerminalMouseDownCapture, handleTerminalMouseUpCapture } = useTerminalClipboard({
+    termRef,
+    containerRef,
+    wsRef,
+    pendingCmdRef,
+    terminalRightClickPasteOnEmptyRef,
+    terminalRightClickPasteModeRef,
+    terminalLeftClickCopyOnSelectionRef,
+    terminalLeftClickCopyOnSelectionModeRef,
+    t,
+  });
   const [timestampsVisible, setTimestampsVisible] = useState(localStorage.getItem('terminalTimestamps') === 'true');
   // 命令块：左侧折叠钮 + 树线，可收起输出
   const commandBlocksEnabledRef = useRef(localStorage.getItem('terminalCommandBlocks') === 'true');
@@ -255,7 +249,6 @@ export default function Terminal({
     tsSnapshotByLine,
     tsRemountFromList,
   });
-  const smartWriteRef = useRef<((data: string | Uint8Array) => void) | null>(null);
 
   // ── 主题/壁纸与设置事件监听（terminal/ 目录同名 hook） ──
   const { T, themeToggle, bgInfo } = useTerminalTheme({ termRef, wrapperRef });
@@ -283,1080 +276,60 @@ export default function Terminal({
     scheduleGutterSync,
   });
 
-  // ── 初始化 xterm + WebSocket 终端通道 ────────────────────────────────
-  // xterm.js 通过 AttachAddon + WebSocket 直接连到本地 Go WebSocket 服务器
-  // 完全绕开 Wails IPC跨进程通信，走 TCP loopback 延迟极低
-  useEffect(() => {
-    if (!containerRef.current) return;
+  // ── xterm/WebSocket 会话主链路（terminal/useTerminalSession） ──
+  useTerminalSession({
+    sessionId,
+    wsRebuildKey,
+    status,
+    isActive,
+    t,
+    T,
+    containerRef,
+    termRef,
+    fitAddonRef,
+    searchAddonRef,
+    wsRef,
+    serverIdRef,
+    shortcutsRef,
+    localEchoRef,
+    timestampsEnabledRef,
+    commandBlocksEnabledRef,
+    alternateBufferActiveRef,
+    setAlternateBufferActive,
+    screenScrollbackRef,
+    prepareScreenScrollbackRef,
+    awaitingPasswordRef,
+    awaitingCommandFinishRef,
+    pendingCmdRef,
+    isTerminalPointerDownRef,
+    dispatchSyntheticTerminalMouseUp,
+    keywordHighlightEnabledRef,
+    hlDecoderRef,
+    hlStateRef,
+    termSearchInputRef,
+    setShowTermSearch,
+    setTermSearchQuery,
+    setTermSearchResult,
+    setContextMenu,
+    setLinkMenu,
+    isShellPromptLine,
+    isCollapseSummaryLine,
+    cbRewriteLockRef,
+    tsSet,
+    tsClearLine,
+    tsClear,
+    cbClear,
+    gutterRef,
+    gutterSyncRAFRef,
+    linkUnderlineLayerRef,
+    linkUnderlineSyncRAFRef,
+    getViewportLinkCache,
+    scheduleGutterSync,
+    scheduleLinkUnderlineSync,
+    handleClearScreen,
+    pasteTerminalSelectionToTerminal,
+  });
 
-    containerRef.current.innerHTML = '';
-
-    const fontSize = parseInt(localStorage.getItem('terminalFontSize') || '13', 10);
-
-    const term = new XTerm({
-      // background 用不透明的容器底色：nano/vim 的反转显示（SGR 7）需要真实的背景色参与
-      // 互换，透明底会退化成黑字黑底；壁纸/色调层改为叠在内容上方保持观感
-      theme:            { ...T.xterm, background: getSolidTerminalBackground(T) },
-      fontFamily:       getResolvedProgramFontPreferences().terminalFontFamily,
-      fontSize:         fontSize,
-      fontWeight:       500,
-      fontWeightBold:   700,
-      lineHeight:       1.22,
-      letterSpacing:    0.3,
-      // 关自动反差：搜索高亮底上白字会被压成黑字
-      minimumContrastRatio: 0,
-      cursorBlink:      true,
-      cursorStyle:      'bar',
-      cursorWidth:      1,
-      scrollback:       5000,
-      // SearchAddon 高亮装饰依赖 proposed API
-      allowProposedApi: true,
-      fastScrollModifier: 'alt',
-      macOptionIsMeta:  true,
-      padding:          8,
-      windowOptions: {
-        setWinSizeChars: true
-      }
-      // xterm 5 类型未声明 padding 选项（运行期仍生效），按构造参数类型断言
-    } as ITerminalOptions & ITerminalInitOnlyOptions & { padding?: number });
-
-    const fitAddon = new FitAddon();
-    term.loadAddon(fitAddon);
-    const searchAddon = new SearchAddon({ highlightLimit: 1000 });
-    term.loadAddon(searchAddon);
-    searchAddonRef.current = searchAddon;
-    const searchResultsDisposable = searchAddon.onDidChangeResults((result) => {
-      setTermSearchResult({
-        resultIndex: typeof result?.resultIndex === 'number' ? result.resultIndex : -1,
-        resultCount: typeof result?.resultCount === 'number' ? result.resultCount : 0,
-      });
-    });
-    // 点击/手型用 provider；常驻下划线用覆盖层。可见区扫描走 getViewportLinkCache
-    const linkProviderDisposable = term.registerLinkProvider({
-      provideLinks(bufferLineNumber, callback) {
-        const found = getViewportLinkCache(term).get(bufferLineNumber) || [];
-        if (!found.length) {
-          callback(undefined);
-          return;
-        }
-        callback(found.map(({ text, range }) => ({
-          text,
-          range,
-          decorations: { underline: false, pointerCursor: true },
-          activate(event, uri) {
-            event?.preventDefault?.();
-            event?.stopPropagation?.();
-            try { term.clearSelection(); } catch (_) {}
-            requestAnimationFrame(() => { try { term.clearSelection(); } catch (_) {} });
-            const x = event?.clientX ?? 0;
-            const y = event?.clientY ?? 0;
-            setContextMenu(null);
-            setLinkMenu({ ...clampMenuPosition(x, y, 200, 96), url: uri });
-          },
-        })));
-      },
-    });
-    term.open(containerRef.current);
-    screenScrollbackRef.current.pending = false;
-    screenScrollbackRef.current.active = false;
-    const syncTuiState = (screenActive = screenScrollbackRef.current.active) => {
-      const active = term.buffer.active.type === 'alternate' || screenActive;
-      alternateBufferActiveRef.current = active;
-      setAlternateBufferActive(active);
-      if (active) {
-        if (gutterSyncRAFRef.current !== null) {
-          cancelAnimationFrame(gutterSyncRAFRef.current);
-          gutterSyncRAFRef.current = null;
-        }
-        if (gutterRef.current) gutterRef.current.innerHTML = '';
-        if (linkUnderlineLayerRef.current) linkUnderlineLayerRef.current.innerHTML = '';
-      } else {
-        scheduleGutterSync();
-        scheduleLinkUnderlineSync();
-      }
-    };
-    prepareScreenScrollbackRef.current = (command) => {
-      if (screenScrollbackRef.current.active) return;
-      screenScrollbackRef.current.pending = startsInteractiveScreen(command);
-    };
-    const screenAltModeSetDisposable = term.parser.registerCsiHandler({ prefix: '?', final: 'h' }, (params) => {
-      const mode = params.length === 1 && typeof params[0] === 'number' ? params[0] : 0;
-      if ((!screenScrollbackRef.current.pending && !screenScrollbackRef.current.active) || (mode !== 47 && mode !== 1047 && mode !== 1049)) return false;
-      screenScrollbackRef.current.pending = false;
-      screenScrollbackRef.current.active = true;
-      syncTuiState(true);
-      return true;
-    });
-    const screenAltModeResetDisposable = term.parser.registerCsiHandler({ prefix: '?', final: 'l' }, (params) => {
-      const mode = params.length === 1 && typeof params[0] === 'number' ? params[0] : 0;
-      if (!screenScrollbackRef.current.active || (mode !== 47 && mode !== 1047 && mode !== 1049)) return false;
-      screenScrollbackRef.current.active = false;
-      screenScrollbackRef.current.pending = false;
-      syncTuiState(false);
-      // normal buffer 已承载 screen 历史，不能再执行 1049l 的旧光标恢复，否则长日志会把提示符拉回已裁剪位置。
-      return true;
-    });
-    try { fitAddon.fit(); } catch (_) {}
-    const terminalInput = containerRef.current.querySelector('.xterm-helper-textarea') as HTMLTextAreaElement | null;
-    if (terminalInput) {
-      terminalInput.name = 'terminalInput';
-      terminalInput.autocomplete = 'off';
-    }
-    alternateBufferActiveRef.current = false;
-    setAlternateBufferActive(false);
-
-    // ── 智能写入：用户手动滚动上时保持位置 ─────────────────────────
-    let userPinned = false; // 用户手动往上滚后锁定
-    const onTermScroll = () => {
-      const buf = term.buffer.active;
-      // 滚到底部时解除锁定
-      if (buf.viewportY >= buf.baseY) {
-        userPinned = false;
-      }
-      scheduleGutterSync();
-      scheduleLinkUnderlineSync();
-    };
-    const scrollDisposable = term.onScroll(onTermScroll);
-    // 直接监听 xterm 视口 DOM scroll 事件作为更可靠的备选
-    const vpEl = containerRef.current.querySelector('.xterm-viewport');
-    if (vpEl) {
-      vpEl.addEventListener('scroll', onTermScroll, { passive: true });
-    }
-
-    // ── 每行时间戳 / 命令块：marker 跟随 xterm scrollback 裁剪 ──
-    const lineFeedDisposable = term.onLineFeed(() => {
-      if (alternateBufferActiveRef.current || term.buffer.active.type !== 'normal') return;
-      if (!timestampsEnabledRef.current && !commandBlocksEnabledRef.current) return;
-
-      const buf = term.buffer.active;
-      const cursorLine = buf.baseY + buf.cursorY;
-      // 往回跳过 isWrapped 包裹行，记到逻辑行首行
-      let pos = cursorLine - 1;
-      while (pos > 0) {
-        const line = buf.getLine(pos);
-        if (line && line.isWrapped) { pos--; } else { break; }
-      }
-      if (pos < 0) return;
-
-      // 收起/展开改写 buffer 时不要打新时间戳；摘要行不打。
-      // 回车完成的行（含「空提示符出现」/「执行命令」）都用当前时刻覆盖旧戳。
-      if (timestampsEnabledRef.current && !cbRewriteLockRef.current) {
-        const posText = buf.getLine(pos)?.translateToString(true) || '';
-        if (!isCollapseSummaryLine(posText)) {
-          tsClearLine(pos);
-          tsSet(term.registerMarker(pos - cursorLine), formatTerminalTimestamp());
-        }
-      }
-      // 命令块由 gutter sync 扫描提示符决定，lineFeed 只需刷新
-      if (commandBlocksEnabledRef.current && !cbRewriteLockRef.current) {
-        scheduleGutterSync();
-      }
-    });
-    const writeParsedDisposable = term.onWriteParsed(() => {
-      scheduleGutterSync();
-      scheduleLinkUnderlineSync();
-      // 命令完成检测：仅当处于"等待命令完成"状态且不在 TUI（备用屏）时，
-      // 检查光标所在行是否已回归 shell 提示符。命中则派发事件并退出等待态。
-      // 提示符识别复用 isShellPromptLine；只取末行避免全量扫描。
-      if (!awaitingCommandFinishRef.current || alternateBufferActiveRef.current) return;
-      const buf = term.buffer.active;
-      if (!buf) return;
-      const lastLine = buf.getLine(buf.baseY + buf.cursorY);
-      const text = lastLine ? lastLine.translateToString(true) : '';
-      if (isShellPromptLine(text)) {
-        awaitingCommandFinishRef.current = false;
-        window.dispatchEvent(new CustomEvent('ssh-command-finished', {
-          detail: { sessionId: serverIdRef.current }
-        }));
-      }
-    });
-    const bufferChangeDisposable = term.buffer.onBufferChange((buffer) => {
-      syncTuiState(buffer.type === 'alternate' || screenScrollbackRef.current.active);
-    });
-    const wheelHandler = (e: WheelEvent) => {
-      // 触控板双指滚动打断选区状态防呆：
-      // 当双指滚动开始时，若由于单指先行接触残留了 isTerminalPointerDownRef 状态，
-      // 且当前无物理按键按下，主动释放状态并闭合选区，防止滚动后单指滑动意外划选文字
-      if (isTerminalPointerDownRef.current && (e.buttons === 0 || !(e.buttons & 1))) {
-        isTerminalPointerDownRef.current = false;
-        dispatchSyntheticTerminalMouseUp(e.clientX, e.clientY);
-      }
-      // 无论向上还是向下滚动，都检查当前位置并更新锁定状态
-      requestAnimationFrame(() => {
-        const buf = term.buffer.active;
-        userPinned = buf.viewportY < buf.baseY;
-      });
-    };
-    containerRef.current?.addEventListener('wheel', wheelHandler, { passive: true });
-
-    const isClearScreenData = (d: string | Uint8Array) => {
-      if (!d) return false;
-      if (typeof d === 'string') return d.includes('\x1b[2J') || d.includes('\x1b[3J');
-      // Binary: scan for \x1b[2J (clear) or \x1b[3J (clear scrollback)
-      if (!d.includes(0x1b)) return false;
-      for (let i = 0; i <= d.length - 4; i++) {
-        if (d[i] === 0x1b && d[i+1] === 0x5b && (d[i+2] === 0x32 || d[i+2] === 0x33) && d[i+3] === 0x4a) {
-          return true;
-        }
-      }
-      return false;
-    };
-    const smartWrite = (data: string | Uint8Array) => {
-      if (isClearScreenData(data)) handleClearScreen();
-      // 关键字高亮：高亮开启时 onmessage 已统一解码为字符串（incomingText），这里只需处理字符串；
-      // 关闭时数据为原始 string/Uint8Array，直接透传不高亮。
-      let writeData = data;
-      if (keywordHighlightEnabledRef.current && typeof data === 'string') {
-        writeData = highlightKeywords(data, hlStateRef.current);
-      }
-      if (userPinned) {
-        // xterm.js 在用户不在底部时已经会保持滚动位置。
-        // 之前用 scrollToLine(savedY) 在异步回调中执行，会在用户向下滚动后
-        // 把视图拉回旧位置，导致用户无法追上最新输出。
-        // 现在仅在 xterm.js 自动滚动打断时才恢复（用相对偏移检测）。
-        const buf = term.buffer.active;
-        const offset = buf.baseY - buf.viewportY;
-        term.write(writeData, () => {
-          const newBuf = term.buffer.active;
-          // 只有当 offset 变小（说明 xterm 自动滚动了）才恢复
-          if (newBuf.baseY - newBuf.viewportY < offset) {
-            const newY = newBuf.baseY - offset;
-            if (newY >= 0) term.scrollToLine(newY);
-          }
-        });
-      } else {
-        term.write(writeData);
-      }
-    };
-    smartWriteRef.current = smartWrite;
-
-    // ── DOM 渲染器（WebGL 在 CJK/宽字符支持差，使用默认 DOM 渲染确保中文正常显示）──
-
-    termRef.current    = term;
-    fitAddonRef.current = fitAddon;
-    window.__luminTerminalSnapshots = window.__luminTerminalSnapshots || {};
-    window.__luminTerminalSnapshots[sessionId] = () => getTerminalBufferSnapshotText(termRef.current || term);
-
-    const fitTimer = setTimeout(() => {
-      try { fitAddon.fit(); } catch (_) {}
-    }, 100);
-
-    // ── 自定义快捷键 ──────────────────────────────────────────────
-
-    // 初始化快捷键缓存（移出按键热路径，仅在首次或变更时读取）
-    if (shortcutsRef.current === null) {
-      let defaults: Record<string, string>;
-      try {
-        const saved = localStorage.getItem('appShortcuts');
-        defaults = saved
-          ? { ...DEFAULT_TERMINAL_SHORTCUTS, ...JSON.parse(saved) }
-          : { ...DEFAULT_TERMINAL_SHORTCUTS };
-      } catch (_) {
-        defaults = { ...DEFAULT_TERMINAL_SHORTCUTS };
-      }
-      shortcutsRef.current = defaults;
-    }
-
-    term.attachCustomKeyEventHandler((e) => {
-      if (e.type !== 'keydown') return true;
-
-      // 修饰键策略：macOS 上 ⌘ = UI 动作（复制/粘贴/清屏/查找），物理 ⌃ = 终端控制信号
-      // （SIGINT/EOF 等，与原生终端一致，⌃V 发送 \x16 literal-next 而非粘贴）；
-      // Win/Linux 上两者合一为 Ctrl。信号/清屏/选区粘贴同时匹配两种组合键。
-
-      // 1. 获取用户自定义的快捷键配置（从 ref 缓存读取，避免热路径访问 localStorage）
-      const customShortcuts: Record<string, string> = shortcutsRef.current || DEFAULT_TERMINAL_SHORTCUTS;
-
-      // 2. 解析当前按下的组合键字符串（macOS 下主快捷键使用 ⌘ Meta，Win/Linux 下使用 Ctrl）
-      const pressedStr = buildCombo(e, getModKey(e));
-      // 3. 基于物理 Ctrl 键的组合键（用于终端控制信号 SIGINT/EOF 等，跨平台始终绑定物理 Ctrl；
-      //    Win/Linux 上与 pressedStr 恒相同，仅 macOS 需要单独构建）
-      const physicalCtrlStr = isMac ? buildCombo(e, e.ctrlKey) : pressedStr;
-
-      // ── 自定义复制键（默认 ⌘C 或 Ctrl+C）：智能处理 ────────
-      if (pressedStr === customShortcuts.copy) {
-        const selection = term.getSelection();
-        if (selection) {
-          e.preventDefault();
-          navigator.clipboard.writeText(selection);
-          term.clearSelection();
-          return false; // 已复制，阻止 xterm 把按键发给服务器
-        }
-        // 【关键】如果没有选区，则直接放行 (return true)
-        // 这样在 Win/Linux 上按 Ctrl+C 能变成标准的终端中断符 (\x03) 发给服务器
-        return true; 
-      }
-
-      // ── Ctrl+Shift+C：强制系统级复制，作为备用方案 ────────
-      if (e.ctrlKey && e.shiftKey && !e.altKey && (e.key === 'c' || e.key === 'C')) {
-        e.preventDefault();
-        const selection = term.getSelection();
-        if (selection) navigator.clipboard.writeText(selection);
-        return false;
-      }
-
-      // ── 自定义粘贴键 ───────────────────────────
-      if (pressedStr === customShortcuts.paste) {
-        // 在 macOS 上按下 ⌘V 且使用默认粘贴配置时，放行给系统原生 paste 事件处理，避免触发 WebKit 异步剪贴板 "Paste" 提示气泡。
-        // 原生路径由 xterm 的 paste 监听器处理：换行归一 + 按需 bracketed paste 包裹（多行粘贴不会逐行自动执行），
-        // 并统一走 term.onData（含 normalizeTerminalPasteText 与本地回显逻辑）。
-        if (isMac && e.metaKey && !e.ctrlKey && !e.shiftKey && !e.altKey && (e.key === 'v' || e.key === 'V') && customShortcuts.paste === 'Ctrl+V') {
-          return true;
-        }
-
-        // 自定义粘贴组合键无法触发系统原生 paste 事件（只有真实的 ⌘V/Ctrl+V 能），
-        // 改走 Wails 运行时剪贴板读取，macOS 下同样不弹 "Paste" 气泡；浏览器 dev 回退 Clipboard API。
-        e.preventDefault();
-        readClipboardText().then((text) => {
-          const payload = normalizeTerminalPasteText(text);
-          if (payload && wsRef.current?.readyState === WebSocket.OPEN) {
-            pendingCmdRef.current += payload.replace(/[\x00-\x1F\x7F]/g, '');
-            wsRef.current.send(textEncoder.encode(payload));
-          }
-        }).catch((err) => {
-          console.error('Clipboard read failed:', err);
-          termRef.current?.focus();
-        });
-        return false;
-      }
-
-      // ── 自定义清屏键 ───────────────────────────
-      if (pressedStr === customShortcuts.clear || physicalCtrlStr === customShortcuts.clear) {
-        e.preventDefault();
-        term.clear();
-        return false;
-      }
-
-      // 新建标签页的快捷键放行给外层 App 处理
-      if (pressedStr === customShortcuts.newTab) {
-        return true;
-      }
-
-      // ── 查找终端缓冲区（默认 ⌘F 或 Ctrl+F）。仅匹配主快捷键：物理 ⌃F 在 macOS 上
-      // 保持终端语义（readline 前进一个字符 \x06），不打开查找 ────────────────
-      const findShortcut = customShortcuts.find || 'Ctrl+F';
-      if (pressedStr === findShortcut) {
-        e.preventDefault();
-        const selection = term.getSelection();
-        setShowTermSearch(true);
-        if (selection && !selection.includes('\n') && selection.length <= 200) {
-          setTermSearchQuery(selection);
-        }
-        requestAnimationFrame(() => {
-          termSearchInputRef.current?.focus();
-          termSearchInputRef.current?.select();
-        });
-        return false;
-      }
-
-      // ── 自定义控制信号（向服务器发送对应的控制字符）。
-      // 同时匹配物理 Ctrl 与主修饰键组合：跨平台物理 Ctrl 始终可用，
-      // macOS 上 ⌘+信号键也生效（保留旧版 ⌘ 映射 Ctrl 的肌肉记忆） ────────────────
-      for (const [key, bytes] of Object.entries(TERMINAL_SIGNAL_BYTES)) {
-        if (customShortcuts[key] && (physicalCtrlStr === customShortcuts[key] || pressedStr === customShortcuts[key])) {
-          e.preventDefault();
-          if (wsRef.current?.readyState === WebSocket.OPEN) {
-            wsRef.current.send(bytes);
-          }
-          return false;
-        }
-      }
-
-      // 已有动作优先，避免重复绑定时一次按键执行两个动作
-      if (pressedStr === customShortcuts.pasteSelection || physicalCtrlStr === customShortcuts.pasteSelection) {
-        e.preventDefault();
-        void pasteTerminalSelectionToTerminal();
-        return false;
-      }
-
-      // ── 其他标准控制字符全部透传给服务器处理 ────────────────────────
-      return true;
-    });
-
-    // ── WebSocket 连接 & Predictive Local Echo ─────────────────────
-    let ws: WebSocket | null = null;
-    let wsConnecting = false;
-    let cancelled = false;
-    const pendingEchoes: string[] = [];
-    let predictiveDecoder = new TextDecoder();
-    let predictiveTextCarry = '';
-    // 重置高亮流式解码器，避免上一次连接的残留字节污染本次输出
-    hlDecoderRef.current = new TextDecoder();
-    // 同步重置前景色状态：上次连接可能在颜色区间未闭合时断开（fgActive=true），
-    // 不清掉会让新连接开局误判为「前景色已激活」而哑火高亮
-    hlStateRef.current = createHighlightState();
-
-    const connectWebSocket = () => {
-      if (cancelled || wsConnecting || statusRef.current !== 'connected' || wsRef.current) return;
-      wsConnecting = true;
-      // 并行获取端口与鉴权 token，后端要求连接时通过 ?token=xxx 携带，防止本机恶意进程注入命令
-      Promise.all([AppGo.GetWsPort(), AppGo.GetWsToken()]).then(([port, token]) => {
-        if (cancelled || statusRef.current !== 'connected' || wsRef.current || !port || !termRef.current) return;
-        const tokenQuery = token ? `?token=${encodeURIComponent(token)}` : '';
-        const currentWs = new WebSocket(`ws://127.0.0.1:${port}/ws/${sessionId}${tokenQuery}`);
-        ws = currentWs;
-        currentWs.binaryType = 'arraybuffer';
-        wsRef.current = currentWs;
-        currentWs.onclose = () => {
-          if (wsRef.current === currentWs) wsRef.current = null;
-        };
-
-      currentWs.onopen = () => {
-        // 补发一次初始尺寸：终端首次 fit 发生在 onResize 订阅之前，那次
-        // 尺寸变化事件被错过，本地 PTY 可能长期停留在出生尺寸；这里主动
-        // 同步一次，同时给 SIGWINCH 会重绘提示符的 shell（bash/zsh）兜底自愈机会。
-        if (termRef.current) {
-          AppGo.ResizeTerminal(sessionId, termRef.current.cols, termRef.current.rows);
-        }
-      };
-
-      currentWs.onmessage = (ev) => {
-        if (!termRef.current) return;
-        // 在原始数据上检测清屏序列（不依赖后续文本处理路径）
-        const rawBytes = typeof ev.data === 'string' ? null : new Uint8Array(ev.data);
-        // 统一解码：高亮开启时整个连接只用一个流式解码器（hlDecoderRef），
-        // 避免「快速路径 / 回显过滤路径」各自持有一个解码器导致跨帧 UTF-8 失步损坏
-        const incomingText = (keywordHighlightEnabledRef.current && rawBytes)
-          ? hlDecoderRef.current.decode(rawBytes, { stream: true })
-          : null;
-        if (timestampsEnabledRef.current) {
-          if (typeof ev.data === 'string' && (ev.data.includes('\x1b[2J') || ev.data.includes('\x1b[3J'))) {
-            handleClearScreen();
-          } else if (rawBytes && rawBytes.includes(0x1b)) {
-            for (let i = 0; i <= rawBytes.length - 4; i++) {
-              if (rawBytes[i] === 0x1b && rawBytes[i+1] === 0x5b && (rawBytes[i+2] === 0x32 || rawBytes[i+2] === 0x33) && rawBytes[i+3] === 0x4a) {
-                handleClearScreen();
-                break;
-              }
-            }
-          }
-        }
-
-          // 检测密码提示，标记下一行输入为密码（不记入命令历史）
-        if (!awaitingPasswordRef.current) {
-          const probeText = incomingText ?? (typeof ev.data === 'string' ? ev.data : textDecoder.decode(ev.data));
-          // ponytail: 只在最后一行像密码/验证码提示时触发（关键词 + 行尾冒号），
-          // 避免 "admin password: xxx" 之类信息性输出误判，导致下一条普通命令被跳过。
-          // 行尾冒号是强约束，关键词可适度放宽：覆盖 OTP/MFA/Token 等验证码提示
-          const lastLine = (probeText.split(/\r?\n/).pop() || '').trim();
-          if (/(password|passwd|passphrase|密码|verification|otp|token|2fa|mfa|auth.*code)/i.test(lastLine) && /[:：]\s*$/.test(lastLine)) {
-            awaitingPasswordRef.current = true;
-          }
-        }
-
-        const shouldFilterIncomingText = (localEchoRef.current && pendingEchoes.length > 0) || predictiveTextCarry.length > 0
-
-        if (!shouldFilterIncomingText) {
-          predictiveDecoder = new TextDecoder()
-          predictiveTextCarry = ''
-          smartWrite(incomingText ?? (typeof ev.data === 'string' ? ev.data : new Uint8Array(ev.data)));
-          return;
-        }
-
-        let text = incomingText ?? (typeof ev.data === 'string' ? ev.data : predictiveDecoder.decode(new Uint8Array(ev.data), { stream: true }));
-        if (predictiveTextCarry) {
-          text = predictiveTextCarry + text;
-          predictiveTextCarry = '';
-        }
-
-        const splitText = splitTrailingIncompleteEscapeSequence(text);
-        predictiveTextCarry = splitText.carry;
-        text = splitText.complete;
-        if (!text) {
-          return;
-        }
-
-        let i = 0;
-        const parts = [];
-
-        while (i < text.length) {
-          // 1. 强大且健壮的 ANSI 转义序列跳过逻辑 (CSI、OSC 及其他单字符转义)
-          if (text[i] === '\x1b') {
-            let j = i + 1;
-            if (j >= text.length) { parts.push(text[i]); i++; continue; }
-            if (text[j] === '[') {
-               // CSI 序列
-               j++;
-               while (j < text.length) {
-                 const c = text.charCodeAt(j);
-                 if (c >= 0x40 && c <= 0x7E) { j++; break; }
-                 j++;
-               }
-            } else if (text[j] === ']') {
-               // OSC 序列 (如 Window Title)
-               j++;
-               while (j < text.length) {
-                 if (text[j] === '\x07') { j++; break; }
-                 if (text[j] === '\x1b' && j + 1 < text.length && text[j+1] === '\\') { j += 2; break; }
-                 j++;
-               }
-            } else {
-               // 其他 ESC 序列（跳过后面一个字符）
-               j++;
-            }
-            parts.push(text.substring(i, j));
-            i = j;
-            continue;
-          }
-
-          // 2. 匹配回显字符并丢弃
-          if (pendingEchoes.length > 0) {
-            const expected = pendingEchoes[0];
-            if (text[i] === expected) {
-              pendingEchoes.shift();
-              i++;
-              continue;
-            }
-            if (expected === '\x7F' && text[i] === '\b') {
-              pendingEchoes.shift();
-              i++;
-              continue;
-            }
-            // 遇到非打印控制字符（如 \r, \n, \x07 等），直接放行打印，不破坏当前的预测队列
-            const charCode = text.charCodeAt(i);
-            if (charCode < 32 || charCode === 127) {
-              parts.push(text[i]);
-              i++;
-              continue;
-            }
-          }
-
-          // 真正的冲突（服务器发来了与预测不符的可打印字符），视为脱轨，清空队列并接受服务器输出
-          pendingEchoes.length = 0;
-          parts.push(text[i]);
-          i++;
-        }
-
-        // 写回经过滤的文本
-        const newText = parts.join('');
-        smartWrite(newText);
-      };
-
-      currentWs.onerror = (e) => console.error('[Terminal] WebSocket error', e);
-      }).finally(() => {
-        wsConnecting = false;
-      });
-    };
-    connectWebSocketRef.current = connectWebSocket;
-    connectWebSocket();
-
-    // ── 历史指令记录 + 输入直觉 + Local Echo ────────────────────────
-    let localInputLength = 0; // 用于保护提示符，防止退格越界
-    let pendingCmdReliable = true;
-
-    term.onData((data) => {
-      if ((statusRef.current === 'closed' || statusRef.current === 'error') && (data.includes('\r') || data.includes('\n'))) {
-        window.dispatchEvent(new CustomEvent('ssh-reconnect-trigger', { detail: sessionId }));
-        return;
-      }
-
-      // 粘贴等多字符输入：把 \\r\\n / \\n 收成单个 \\r，保证 bash 的 \\ 续行不断
-      let out = data;
-      if (out.length > 1 && /[\r\n]/.test(out)) {
-        out = normalizeTerminalPasteText(out);
-      }
-
-      if (wsRef.current?.readyState === WebSocket.OPEN) {
-        wsRef.current.send(textEncoder.encode(out));
-      }
-
-      // ── 命令记录：优先读取终端可见缓冲区的当前行（屏幕上实际渲染、
-      // 真正被 shell 执行的命令），可正确捕捉 Tab 补全 / 方向键调历史 /
-      // Ctrl+R 等 shell 编辑结果；pendingCmdRef 仅在缓冲读取失败时作兜底。
-      if (out.includes('\r') || out.includes('\n')) {
-        if (term.buffer.active.type === 'alternate') {
-          pendingCmdRef.current = '';
-          return;
-        }
-        // 多行粘贴：只把最后一行之前的可见内容并入历史，避免把整段 paste 拆烂
-        const lines = out.split(/\r/).filter((line, i, arr) => i < arr.length - 1 || line.length > 0);
-        if (lines.length > 1) {
-          for (const line of lines) {
-            const piece = line.replace(/[\x00-\x1F\x7F]/g, '');
-            if (piece) pendingCmdRef.current += (pendingCmdRef.current ? ' ' : '') + piece;
-          }
-        } else {
-          const nlIdx = out.search(/[\r\n]/);
-          if (nlIdx > 0) {
-            pendingCmdRef.current += out.slice(0, nlIdx).replace(/[\x00-\x1F\x7F]/g, '');
-          }
-        }
-        let cmd = '';
-        const buf = term.buffer.active;
-        let promptFilteredThisTurn = false;
-        if (buf) {
-          const bufLine = buf.getLine(buf.baseY + buf.cursorY);
-          const text = bufLine ? bufLine.translateToString(true) : '';
-          cmd = extractCommandFromBufferLine(text);
-          // 含控制字符（C0 0x00-0x1F / DEL / C1 0x80-0x9F，多为 ANSI 序列残留）
-          // 或交互脚本提示：视为无效，回退到逐字符累加。注意保留合法 Unicode
-          // （如提示符 ❯、中文路径、emoji 参数），不再按"非 ASCII"一刀切丢弃。
-          const hasControl = /[\x00-\x1F\x7F-\x9F]/.test(cmd);
-          const isPrompt = isInteractivePromptText(cmd);
-          if (hasControl || isPrompt) {
-            promptFilteredThisTurn = true;
-            cmd = '';
-          }
-        }
-        const pending = pendingCmdRef.current.trim();
-        if (!promptFilteredThisTurn) {
-          if (!cmd) {
-            cmd = pending;
-          } else if (pendingCmdReliable && pending) {
-            const c = cmd.toLowerCase(), p = pending.toLowerCase();
-            if (!c.startsWith(p) && !p.startsWith(c)) cmd = '';
-          }
-        }
-        if (!awaitingPasswordRef.current) {
-          prepareScreenScrollbackRef.current(cmd);
-        }
-        if (!awaitingPasswordRef.current && cmd.length > 1 && !/^\d+$/.test(cmd)) {
-          window.dispatchEvent(new CustomEvent('ssh-command-history', {
-            detail: { sessionId: serverIdRef.current, command: cmd, time: new Date().toISOString(), source: 'input' }
-          }));
-        }
-        // 非密码输入且提交了实际命令：进入"等待命令完成"状态，
-        // 待提示符回归（onWriteParsed 检测）时派发 ssh-command-finished 事件，
-        // 供文件管理器自动刷新当前目录。
-        awaitingCommandFinishRef.current = !awaitingPasswordRef.current && cmd.length > 0;
-        awaitingPasswordRef.current = false;
-        pendingCmdRef.current = '';
-        pendingCmdReliable = true;
-      } else if (out === '\x7F' || out === '\b') {
-        pendingCmdRef.current = pendingCmdRef.current.slice(0, -1);
-      } else if (!/[\x00-\x1F\x7F]/.test(out)) {
-        pendingCmdRef.current += out;
-      } else if (out === '\x03' || out === '\x04') {
-        pendingCmdRef.current = '';
-        pendingCmdReliable = true;
-        if (!screenScrollbackRef.current.active) screenScrollbackRef.current.pending = false;
-        awaitingPasswordRef.current = false; // Ctrl+C/D 取消当前输入，重置密码等待状态，避免下一条普通命令被误跳过
-      } else {
-        pendingCmdReliable = false;
-      }
-
-      // Local Echo 逻辑 (恢复默认开启)
-      if (localEchoRef.current) {
-        // 如果输入中不包含控制字符（如方向键、Esc、退格等），则视作常规可见输入（支持多字符连击或粘贴）
-        if (!/[\x00-\x1F\x7F]/.test(out)) {
-          // 由于 JavaScript 中部分多字节字符的 length 表现，这里按照字符串常规长度累加是安全的。
-          // 因为退格也是按字符来删的。
-          localInputLength += out.length;
-          for (let i = 0; i < out.length; i++) {
-            pendingEchoes.push(out[i]);
-          }
-          term.write(out);
-        } else if (out === '\x7F') { // Backspace
-          // 仅当我们确信这是用户刚刚输入的字符时，才在本地执行退格预测。
-          // 否则（localInputLength <= 0），将退格完全交还给服务器，保护提示符不被删除。
-          if (localInputLength > 0) {
-            localInputLength--;
-            pendingEchoes.push(out);
-            term.write('\b \b'); // 本地立即执行退格效果
-          }
-        } else if (out === '\r' || out === '\n' || out === '\r\n' || (out.length > 1 && /[\r\n]/.test(out))) {
-          localInputLength = 0;
-        } else {
-          // 遇到方向键、Ctrl快捷键（如 Ctrl+C/D/Z）等控制符，
-          // 立刻清零预测输入长度，安全退回到服务器渲染模式
-          localInputLength = 0;
-        }
-      }
-
-    });
-
-    const resizeDisposable = term.onResize(({ cols, rows }) => {
-      AppGo.ResizeTerminal(sessionId, cols, rows);
-      scheduleGutterSync();
-      scheduleLinkUnderlineSync();
-    });
-    // 首帧同步常驻下划线
-    scheduleLinkUnderlineSync();
-
-    return () => {
-      cancelled = true;
-      if (connectWebSocketRef.current === connectWebSocket) connectWebSocketRef.current = null;
-      scrollDisposable.dispose();
-      lineFeedDisposable.dispose();
-      writeParsedDisposable.dispose();
-      bufferChangeDisposable.dispose();
-      screenAltModeSetDisposable.dispose();
-      screenAltModeResetDisposable.dispose();
-      resizeDisposable.dispose();
-      try { linkProviderDisposable.dispose(); } catch (_) {}
-      try { searchResultsDisposable.dispose(); } catch (_) {}
-      try { searchAddon.dispose(); } catch (_) {}
-      if (gutterSyncRAFRef.current !== null) {
-        cancelAnimationFrame(gutterSyncRAFRef.current);
-        gutterSyncRAFRef.current = null;
-      }
-      if (linkUnderlineSyncRAFRef.current !== null) {
-        cancelAnimationFrame(linkUnderlineSyncRAFRef.current);
-        linkUnderlineSyncRAFRef.current = null;
-      }
-      if (linkUnderlineLayerRef.current) linkUnderlineLayerRef.current.innerHTML = '';
-      clearTimeout(fitTimer);
-      if (vpEl) vpEl.removeEventListener('scroll', onTermScroll);
-      // 移除 wheel 监听器，避免内存泄漏
-      containerRef.current?.removeEventListener('wheel', wheelHandler);
-      if (ws) { try { ws.close(); } catch (_) {} }
-      if (wsRef.current === ws) wsRef.current = null;
-      tsClear(); // 清理时间戳
-      cbClear(); // 清理命令块边框
-      if (window.__luminTerminalSnapshots?.[sessionId]) {
-        delete window.__luminTerminalSnapshots[sessionId];
-      }
-      smartWriteRef.current = null;
-      screenScrollbackRef.current.pending = false;
-      screenScrollbackRef.current.active = false;
-      prepareScreenScrollbackRef.current = () => {};
-      alternateBufferActiveRef.current = false;
-      termRef.current     = null;
-      fitAddonRef.current = null;
-      searchAddonRef.current = null;
-      // xterm Viewport 构造里 setTimeout(syncScrollArea) 无句柄；StrictMode 先 dispose 再触发会读空 renderer.dimensions
-      // 延后 dispose，让该 setTimeout 先跑完（同队列 FIFO）
-      const termToDispose = term;
-      setTimeout(() => {
-        try { termToDispose.dispose(); } catch (_) {}
-      }, 0);
-    };
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [sessionId, wsRebuildKey]);
-
-  // ── 监听字体大小修改事件 ──────────────────────────────────────
-  useEffect(() => {
-    const handleFontSizeChange = (e: Event) => {
-      if (termRef.current) {
-        termRef.current.options.fontSize = (e as CustomEvent<number>).detail;
-        if (fitAddonRef.current) {
-          try { fitAddonRef.current.fit(); } catch (_) {}
-        }
-        scheduleGutterSync();
-      }
-    };
-    window.addEventListener('terminal-font-size-changed', handleFontSizeChange);
-    return () => window.removeEventListener('terminal-font-size-changed', handleFontSizeChange);
-  }, []);
-
-  // SSH/本地/串口断开时终端组件会保活以保留输出；单独关闭 WS，释放
-  // 浏览器连接和 Go ReadMessage goroutine。重连后只重建 WS，不重建 xterm。
-  useEffect(() => {
-    if (status === 'closed' || status === 'error') {
-      screenScrollbackRef.current.pending = false;
-      screenScrollbackRef.current.active = false;
-      const actualAlternate = termRef.current?.buffer.active.type === 'alternate';
-      alternateBufferActiveRef.current = actualAlternate;
-      setAlternateBufferActive(actualAlternate);
-      const ws = wsRef.current;
-      wsRef.current = null;
-      if (ws) {
-        try { ws.close(); } catch (_) {}
-      }
-      return;
-    }
-    if (status === 'connected') {
-      connectWebSocketRef.current?.();
-    }
-  }, [status]);
-
-  // ── 状态变化提示 ─────────────────────────────────────────────────
-  useEffect(() => {
-    if (!termRef.current) return;
-    const sw = smartWriteRef.current;
-    if (status === 'error') {
-      sw ? sw('\r\n\x1b[31m✗ ' + t('连接失败') + '\x1b[0m\r\n') : termRef.current.write('\r\n\x1b[31m✗ ' + t('连接失败') + '\x1b[0m\r\n');
-    } else if (status === 'closed') {
-      sw ? sw('\r\n\x1b[33m⚠ ' + t('已断开') + '\x1b[0m\r\n') : termRef.current.write('\r\n\x1b[33m⚠ ' + t('已断开') + '\x1b[0m\r\n');
-    }
-  }, [status]);
-
-  // ── 监听容器大小变化进行自适应 ───────────────────────────────────
-  useEffect(() => {
-    if (!isActive || !containerRef.current || !fitAddonRef.current || !termRef.current) return;
-
-    let resizeTimer: ReturnType<typeof setTimeout> | null = null;
-    const observer = new ResizeObserver((entries) => {
-      if (resizeTimer) clearTimeout(resizeTimer);
-      resizeTimer = setTimeout(() => {
-        if (!termRef.current || !fitAddonRef.current || !containerRef.current) return;
-        const rect = containerRef.current.getBoundingClientRect();
-        if (rect.width <= 0 || rect.height <= 0) return;
-        try {
-          fitAddonRef.current.fit();
-          const { cols, rows } = termRef.current;
-          AppGo.ResizeTerminal(sessionId, cols, rows);
-        } catch (e) {
-          console.error('[Terminal] Resize error:', e);
-        }
-      }, 50);
-    });
-
-    observer.observe(containerRef.current);
-
-    return () => {
-      if (resizeTimer) clearTimeout(resizeTimer);
-      observer.disconnect();
-    };
-  }, [isActive, sessionId]);
-
-  // ── 终端切换回来时，重新 fit ────────────────────────────────────
-  useEffect(() => {
-    if (!isActive || !termRef.current || !fitAddonRef.current) return;
-    const term = termRef.current;
-    const fitAddon = fitAddonRef.current;
-    const raf = requestAnimationFrame(() => {
-      try {
-        const rect = containerRef.current?.getBoundingClientRect();
-        if (rect && rect.width > 0 && rect.height > 0) {
-          fitAddon.fit();
-          const { cols, rows } = term;
-          AppGo.ResizeTerminal(sessionId, cols, rows);
-        }
-      } catch (e) {
-        console.error('[Terminal] activate fit error:', e);
-      }
-    });
-    return () => cancelAnimationFrame(raf);
-  }, [isActive, sessionId]);
-
-  const getTerminalBufferCellPositionFromMouseEvent = useCallback((event: React.MouseEvent, isSelection = false) => {
-    const term = termRef.current;
-    const container = containerRef.current;
-    if (!term?.buffer?.active || !container || typeof window === 'undefined') {
-      return null;
-    }
-    const screen = container.querySelector('.xterm-screen');
-    if (!screen) {
-      return null;
-    }
-    const rect = screen.getBoundingClientRect();
-    if (!rect.width || !rect.height || !term.cols || !term.rows) {
-      return null;
-    }
-    const style = window.getComputedStyle(screen);
-    const leftPadding = parseInt(style.getPropertyValue('padding-left'), 10) || 0;
-    const topPadding = parseInt(style.getPropertyValue('padding-top'), 10) || 0;
-    const cellWidth = rect.width / term.cols;
-    const cellHeight = rect.height / term.rows;
-    if (!Number.isFinite(cellWidth) || !Number.isFinite(cellHeight) || cellWidth <= 0 || cellHeight <= 0) {
-      return null;
-    }
-    const relativeX = event.clientX - rect.left - leftPadding;
-    const relativeY = event.clientY - rect.top - topPadding;
-    let x = Math.ceil((relativeX + (isSelection ? cellWidth / 2 : 0)) / cellWidth);
-    let viewportRow = Math.ceil(relativeY / cellHeight);
-    x = Math.min(Math.max(x, 1), term.cols + (isSelection ? 1 : 0)) - 1;
-    viewportRow = Math.min(Math.max(viewportRow, 1), term.rows) - 1;
-    return {
-      x,
-      y: term.buffer.active.viewportY + viewportRow,
-    };
-  }, []);
-
-  const isTerminalBufferCellWithinRange = useCallback((position: { x: number; y: number } | null, range: IBufferRange | null | undefined) => {
-    if (!position || !range?.start || !range?.end) {
-      return false;
-    }
-    return (position.y > range.start.y && position.y < range.end.y)
-      || (range.start.y === range.end.y && position.y === range.start.y && position.x >= range.start.x && position.x < range.end.x)
-      || (range.start.y < range.end.y && position.y === range.end.y && position.x < range.end.x)
-      || (range.start.y < range.end.y && position.y === range.start.y && position.x >= range.start.x);
-  }, []);
-
-  const copyTerminalSelectionText = useCallback((text: string) => {
-    if (!text) {
-      return;
-    }
-    navigator.clipboard.writeText(text).then(() => {
-      termRef.current?.focus();
-    }).catch((err) => {
-      console.error('Failed to write clipboard:', err);
-      termRef.current?.focus();
-    });
-  }, []);
-
-  const pasteClipboardToTerminal = useCallback(() => {
-    // 走 Wails 运行时读取剪贴板：右键/菜单粘贴没有 keydown 可放行成原生 paste 事件，
-    // macOS 下 navigator.clipboard.readText() 会弹 "Paste" 提示气泡（issue #263）
-    readClipboardText().then((text) => {
-      const payload = normalizeTerminalPasteText(text);
-      if (payload && wsRef.current && wsRef.current.readyState === WebSocket.OPEN) {
-        pendingCmdRef.current += payload.replace(/[\x00-\x1F\x7F]/g, '');
-        wsRef.current.send(textEncoder.encode(payload));
-      }
-      termRef.current?.focus();
-    }).catch((err) => {
-      console.error('Failed to read clipboard:', err);
-      termRef.current?.focus();
-    });
-  }, []);
-
-  const pasteTerminalSelectionToTerminal = useCallback(async () => {
-    const term = termRef.current;
-    const selectedText = term?.getSelection?.() || '';
-    if (!selectedText || !term) {
-      term?.focus();
-      return;
-    }
-
-    const lineCount = selectedText.replace(/\r\n?/g, '\n').split('\n').length;
-    if (lineCount > 3 && localStorage.getItem('skipTerminalSelectionPasteConfirm') !== 'true') {
-      const result = await window.luminDialog?.confirm(
-        t('所选内容超过3行，是否继续粘贴？'),
-        t('确认粘贴'),
-        t('不再询问')
-      );
-      const confirmed = typeof result === 'object' ? result.confirmed : result === true;
-      if (!confirmed) {
-        term.focus();
-        return;
-      }
-      if (typeof result === 'object' && result.checked) {
-        localStorage.setItem('skipTerminalSelectionPasteConfirm', 'true');
-      }
-    }
-
-    const payload = normalizeTerminalPasteText(selectedText);
-    if (payload && wsRef.current?.readyState === WebSocket.OPEN) {
-      pendingCmdRef.current += payload.replace(/[\x00-\x1F\x7F]/g, '');
-      wsRef.current.send(textEncoder.encode(payload));
-      term.clearSelection();
-    }
-    term.focus();
-  }, [t]);
-
-  const handleTerminalMouseDownCapture = useCallback((event: React.MouseEvent) => {
-    if (event.button === 0) {
-      isTerminalPointerDownRef.current = true;
-    }
-    if (event.button !== 0 || !terminalLeftClickCopyOnSelectionRef.current) {
-      terminalMouseDownSelectionRef.current = null;
-      return;
-    }
-    const mode = terminalLeftClickCopyOnSelectionModeRef.current === 'mouseup' ? 'mouseup' : 'click';
-    if (mode === 'mouseup') {
-      terminalMouseDownSelectionRef.current = {
-        mode,
-        startClientX: event.clientX,
-        startClientY: event.clientY,
-      };
-      return;
-    }
-    const term = termRef.current;
-    const text = term?.getSelection?.() || '';
-    const range = term?.getSelectionPosition?.();
-    const position = getTerminalBufferCellPositionFromMouseEvent(event, true);
-    if (!text || !range || !position || !isTerminalBufferCellWithinRange(position, range)) {
-      terminalMouseDownSelectionRef.current = null;
-      return;
-    }
-    terminalMouseDownSelectionRef.current = {
-      mode,
-      text,
-      startClientX: event.clientX,
-      startClientY: event.clientY,
-    };
-  }, [getTerminalBufferCellPositionFromMouseEvent, isTerminalBufferCellWithinRange]);
-
-  const handleTerminalMouseUpCapture = useCallback((event: React.MouseEvent) => {
-    if (event.button === 0) {
-      isTerminalPointerDownRef.current = false;
-    }
-    const snapshot = terminalMouseDownSelectionRef.current;
-    terminalMouseDownSelectionRef.current = null;
-    if (event.button !== 0 || !terminalLeftClickCopyOnSelectionRef.current || !snapshot) {
-      return;
-    }
-    const deltaX = Math.abs(event.clientX - snapshot.startClientX);
-    const deltaY = Math.abs(event.clientY - snapshot.startClientY);
-    if (snapshot.mode === 'mouseup') {
-      if (deltaX <= 4 && deltaY <= 4) {
-        return;
-      }
-      requestAnimationFrame(() => {
-        const text = termRef.current?.getSelection?.() || '';
-        if (!text) {
-          return;
-        }
-        copyTerminalSelectionText(text);
-      });
-      return;
-    }
-    if (deltaX > 4 || deltaY > 4) {
-      return;
-    }
-    event.preventDefault();
-    event.stopPropagation();
-    copyTerminalSelectionText(snapshot.text ?? '');
-  }, [copyTerminalSelectionText]);
-
-  useEffect(() => {
-    const handleWindowMouseMove = (event: MouseEvent) => {
-      // macOS WKWebView / 触控板手势丢失 mouseup 防呆：
-      // 当物理上没有按键处于按下状态（buttons === 0 或未按左键），但终端仍记录着按下状态时，
-      // 说明上一次 mousedown 对应的 mouseup 已丢失。主动派发合成 mouseup 闭合 xterm 拖拽选区状态机。
-      if (isTerminalPointerDownRef.current && (event.buttons === 0 || !(event.buttons & 1))) {
-        isTerminalPointerDownRef.current = false;
-        dispatchSyntheticTerminalMouseUp(event.clientX, event.clientY);
-      }
-    };
-
-    const handleWindowMouseUp = (event: MouseEvent) => {
-      if (event.button === 0) {
-        isTerminalPointerDownRef.current = false;
-      }
-      const snapshot = terminalMouseDownSelectionRef.current;
-      if (event.button !== 0 || !terminalLeftClickCopyOnSelectionRef.current || !snapshot || snapshot.mode !== 'mouseup') {
-        return;
-      }
-      terminalMouseDownSelectionRef.current = null;
-      const deltaX = Math.abs(event.clientX - snapshot.startClientX);
-      const deltaY = Math.abs(event.clientY - snapshot.startClientY);
-      if (deltaX <= 4 && deltaY <= 4) {
-        return;
-      }
-      requestAnimationFrame(() => {
-        const text = termRef.current?.getSelection?.() || '';
-        if (!text) {
-          return;
-        }
-        copyTerminalSelectionText(text);
-      });
-    };
-
-    const handleWindowPointerCancel = () => {
-      if (isTerminalPointerDownRef.current) {
-        isTerminalPointerDownRef.current = false;
-        dispatchSyntheticTerminalMouseUp();
-      }
-    };
-
-    const handleWindowBlur = () => {
-      // 失焦常意味着按键最终在 WebView 之外释放（mouseup 不会送达），xterm 仍卡在拖选态。
-      // 若只清 isTerminalPointerDownRef，检测基准被清空后这个卡死态将永远无法被发现，划动又会拖选；
-      // 因此失焦时同样派发合成 mouseup 闭合它。快照需先清，避免合成 mouseup 冒泡触发 mouseup 模式的复制。
-      const wasPointerDown = isTerminalPointerDownRef.current;
-      isTerminalPointerDownRef.current = false;
-      terminalMouseDownSelectionRef.current = null;
-      if (wasPointerDown) {
-        dispatchSyntheticTerminalMouseUp();
-      }
-    };
-
-    window.addEventListener('mousemove', handleWindowMouseMove, { capture: true, passive: true });
-    window.addEventListener('mouseup', handleWindowMouseUp);
-    window.addEventListener('pointercancel', handleWindowPointerCancel);
-    window.addEventListener('dragend', handleWindowPointerCancel);
-    window.addEventListener('blur', handleWindowBlur);
-    return () => {
-      window.removeEventListener('mousemove', handleWindowMouseMove, true);
-      window.removeEventListener('mouseup', handleWindowMouseUp);
-      window.removeEventListener('pointercancel', handleWindowPointerCancel);
-      window.removeEventListener('dragend', handleWindowPointerCancel);
-      window.removeEventListener('blur', handleWindowBlur);
-    };
-  }, [copyTerminalSelectionText]);
 
   const handleContextMenu = (e: React.MouseEvent) => {
     e.preventDefault();
